+++ R/bar.R
@@ -14,6 +14,7 @@
 #'       barUI("bar")
 #'     ),
 #'     mainPanel(
+#'       optionUI("bar"),
 #'       plotOutput("bar_plot"),
 #'       ggplotdownUI("bar")
 #'     )
@@ -55,7 +56,6 @@ barUI <- function(id, label = "barplot") {
 }
 
 
-
 optionUI <- function(id) {
   # Create a namespace function using the provided id
   ns <- NS(id)
@@ -87,6 +87,7 @@ optionUI <- function(id) {
 #'       barUI("bar")
 #'     ),
 #'     mainPanel(
+#'       optionUI("bar"),
 #'       plotOutput("bar_plot"),
 #'       ggplotdownUI("bar")
 #'     )
@@ -110,10 +111,12 @@ optionUI <- function(id) {
 #' @export
 #' @import shiny
 #' @importFrom data.table data.table .SD :=
-#' @importFrom ggpubr ggbarplot
+#' @importFrom ggpubr ggbarplot stat_compare_means geom_pwc
 #' @importFrom ggplot2 ggsave
 #' @importFrom rvg dml
 #' @importFrom officer read_pptx add_slide ph_with ph_location
+#' @importFrom scales label_pvalue
+#' @importFrom shinyWidgets dropdownButton tooltipOptions
 
 
 
@@ -203,7 +206,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
             ),
             tabPanel(
               "strataFalse",
-              checkboxInput(session$ns("isPvalue"), "P value?"),
+              checkboxInput(session$ns("isPvalue"), "P value"),
             )
           ),
           tabsetPanel(
@@ -234,7 +237,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
             ),
             tabPanel(
               "over_three",
-              checkboxInput(session$ns("isPair"), "Pair sample P value?"),
+              checkboxInput(session$ns("isPair"), "Pairwise P value"),
             )
           ),
           tabsetPanel(
@@ -261,7 +264,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
             selected = "strataFalse",
             tabPanel(
               "strataTrue",
-              checkboxInput(session$ns("isStrata"), "Pair sample P value?"),
+              checkboxInput(session$ns("isStrata"), "Pairwise P value"),
             ),
             tabPanel(
               "strataFalse",
@@ -498,7 +501,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
 
         if (input$isPvalue & input$strata == "None") {
           res.plot <- res.plot +
-            stat_compare_means(
+            ggpubr::stat_compare_means(
               method = pval.name,
               size = pval.font.size[1],
               label.x.npc = pval.coord[1],
@@ -511,7 +514,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
 
         if (input$isPair & vlist()$nclass_factor[input$x_bar] > 2 & input$strata == "None") {
           res.plot <- res.plot +
-            geom_pwc(
+            ggpubr::geom_pwc(
               method = ppval.name,
               size = pval.font.size[3],
               label.size = pval.font.size[2],
@@ -521,7 +524,7 @@ barServer <- function(id, data, data_label, data_varStruct = NULL, nfactor.limit
 
         if (input$isStrata & input$strata != "None") {
           res.plot <- res.plot +
-            geom_pwc(
+            ggpubr::geom_pwc(
               method = spval.name,
               size = pval.font.size[3],
               label.size = pval.font.size[2],